auth raglasha

+++ app/Http/Livewire/SoldPolicyShow.php
@@ -494,6 +494,7 @@ class SoldPolicyShow extends Component
 
     public function updatedCommDoc()
     {
+        $this->authorize('update',SalesComm::find($this->commDocId));
         $this->validate([
             'commDoc' => 'required|file|mimes:pdf,doc,docx,xls,xlsx,jpg,jpeg,png,bmp,gif,svg,webp|max:5120',
         ]);
@@ -540,6 +541,7 @@ class SoldPolicyShow extends Component
 
     public function removeCommDoc()
     {
+        $this->authorize('update',SalesComm::find($this->RemoveCommDocId));
         $res = SalesComm::find($this->RemoveCommDocId)->deleteDocument();
         if ($res) {
             $this->mount($this->soldPolicy->id);
@@ -552,6 +554,7 @@ class SoldPolicyShow extends Component
 
     public function setCommPaid($id)
     {
+        $this->authorize('update',SalesComm::find($id));
         $res =  SalesComm::find($id)->setAsPaid();
         if ($res) {
             $this->mount($this->soldPolicy->id);
@@ -563,6 +566,7 @@ class SoldPolicyShow extends Component
 
     public function setCommCancelled($id)
     {
+        $this->authorize('update',SalesComm::find($id));
         $res =  SalesComm::find($id)->setAsCancelled();
         if ($res) {
             $this->mount($this->soldPolicy->id);
@@ -574,6 +578,7 @@ class SoldPolicyShow extends Component
 
     public function refreshCommAmmount($id)
     {
+        $this->authorize('update',SalesComm::find($id));
         $res =  SalesComm::find($id)->refreshPaymentInfo();
         if ($res) {
             $this->mount($this->soldPolicy->id);
